moved tests to the autocomplete class
